+++ spec/javascripts/govuk_publishing_components/analytics-ga4/ga4-page-views.spec.js
@@ -44,7 +44,9 @@ describe('Google Tag Manager page view tracking', function () {
         political_status: undefined,
         primary_publishing_organisation: undefined,
         organisations: undefined,
-        world_locations: undefined
+        world_locations: undefined,
+
+        dynamic: 'false'
       }
     }
     window.dataLayer = []
@@ -309,5 +311,22 @@ describe('Google Tag Manager page view tracking', function () {
     linkForURLMock.click()
 
     expect(window.dataLayer[0]).toEqual(expected)
+  })
+
+  describe('correctly sets the referrer and dynamic properties', function () {
+    it('when not passed a parameter', function () {
+      GOVUK.analyticsGa4.analyticsModules.PageViewTracker.init()
+
+      expect(window.dataLayer[0]).toEqual(expected)
+    })
+
+    it('when passed a parameter for the referrer', function () {
+      expected.page_view.referrer = 'https://gov.uk/referrer'
+      expected.page_view.dynamic = 'true'
+
+      GOVUK.analyticsGa4.analyticsModules.PageViewTracker.init('https://gov.uk/referrer')
+
+      expect(window.dataLayer[0]).toEqual(expected)
+    })
   })
 })