+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Fix issue with blue action link arrow svg ([PR #3039](https://github.com/alphagov/govuk_publishing_components/pull/3039))
 * **BREAKING:** Fix referrer bug ([PR #3032](https://github.com/alphagov/govuk_publishing_components/pull/3032))
+* Remove target from mailto: link ([PR #3023](https://github.com/alphagov/govuk_publishing_components/pull/3023))
 
 ## 31.2.0